revert and remove CommandScopes
