+++ edge/op.conf
@@ -1,0 +1,2 @@
+build: docker build -t dannyben/bashly:edge . && docker images |grep bashly
+push: docker push dannyben/bashly:edge
