+++ lib/bashly/concerns/renderable.rb
@@ -18,6 +18,22 @@ module Bashly
       "# #{id}" unless Settings.production?
     end
 
+    # Reads a file from the userspace (Settings.source_dir) and returns
+    # its contents. If the file is not found, returns a string with a hint.
+    def load_user_file(file, placeholder: true)
+      path = "#{Settings.source_dir}/#{file}"
+
+      content = if File.exist? path
+        File.read(path).remove_front_matter
+      elsif placeholder
+        %q[echo "error: cannot load file"]
+      else
+        ''
+      end
+
+      Settings.production? ? content : "#{view_marker path}\n#{content}"
+    end
+
   private
 
     def view_path(view)
