+++ lib/bashly/script/command.rb
@@ -2,7 +2,6 @@ module Bashly
   module Script
     class Command < Base
       include Completions::Command
-      include CommandScopes
 
       class << self
         def option_keys
@@ -32,6 +31,16 @@ module Bashly
         [name] + alt
       end
 
+      # Returns an array of all full names (including aliases and aliases of
+      # parents)
+      def all_full_names
+        if parent_command
+          parent_command.all_full_names.product(aliases).map { |a| a.join ' ' }
+        else
+          aliases
+        end
+      end
+
       # Returns an array of alternative aliases if any
       def alt
         # DEPRECATION 0.8.0
@@ -57,6 +66,33 @@ module Bashly
         @catch_all ||= CatchAll.from_config options['catch_all']
       end
 
+      # Returns a full list of the Command names and aliases combined
+      def command_aliases
+        commands.map(&:aliases).flatten
+      end
+
+      # Returns a data structure for displaying subcommands help
+      def command_help_data
+        result = {}
+
+        public_commands.each do |command|
+          result[command.group_string] ||= {}
+          result[command.group_string][command.name] = command.summary_string
+          next unless command.expose
+
+          command.public_commands.each do |subcommand|
+            result[command.group_string]["#{command.name} #{subcommand.name}"] = subcommand.summary_string
+          end
+        end
+
+        result
+      end
+
+      # Returns only the names of the Commands
+      def command_names
+        commands.map &:name
+      end
+
       # Returns an array of the Commands
       def commands
         return [] unless options["commands"]
@@ -65,6 +101,40 @@ module Bashly
           options['parent_command'] = self
           Command.new options
         end
+      end
+
+      # Returns a flat array containing all the commands in this tree.
+      # This includes self + children + grandchildres + ...
+      def deep_commands
+        result = []
+        commands.each do |command|
+          result << command
+          if command.commands.any?
+            result += command.deep_commands
+          end
+        end
+        result
+      end
+
+      # If any of this command's subcommands has the default option set to
+      # true, this default command will be returned, nil otherwise.
+      def default_command
+        commands.find { |c| c.default }
+      end
+
+      # Returns an array of all the default Args
+      def default_args
+        args.select &:default
+      end
+
+      # Returns an array of all the default Environment Variables
+      def default_environment_variables
+        environment_variables.select &:default
+      end
+
+      # Returns an array of all the default Flags
+      def default_flags
+        flags.select &:default
       end
 
       # Returns an array of EnvironmentVariables
@@ -109,23 +179,6 @@ module Bashly
         end
       end
 
-      # Reads a file from the userspace (Settings.source_dir) and returns
-      # its contents. 
-      # If the file is not found, returns a string with a hint.
-      def load_user_file(file, placeholder: true)
-        path = "#{Settings.source_dir}/#{file}"
-
-        content = if File.exist? path
-          File.read(path).remove_front_matter
-        elsif placeholder
-          %q[echo "error: cannot load file"]
-        else
-          ''
-        end
-
-        Settings.production? ? content : "#{view_marker path}\n#{content}"
-      end
-
       # Returns the Command instance of the direct parent
       def parent_command
         options['parent_command']
@@ -137,9 +190,29 @@ module Bashly
         options['parents'] || []
       end
 
+      # Returns only commands that are not private
+      def public_commands
+        commands.reject &:private
+      end
+
       # Returns true if one of the args is repeatable
       def repeatable_arg_exist?
         args.select(&:repeatable).any?
+      end
+
+      # Returns an array of all the required Arguments
+      def required_args
+        args.select &:required
+      end
+
+      # Returns an array of all the required EnvironmentVariables
+      def required_environment_variables
+        environment_variables.select &:required
+      end
+
+      # Returns an array of all the required Flags
+      def required_flags
+        flags.select &:required
       end
 
       # Returns true if this is the root command (no parents)
@@ -180,6 +253,15 @@ module Bashly
         @user_lib ||= Dir["#{Settings.full_lib_dir}/**/*.sh"].sort
       end
 
+      # Returns an array of all the args with a whitelist
+      def whitelisted_args
+        args.select &:allowed
+      end
+
+      # Returns an array of all the flags with a whitelist arg
+      def whitelisted_flags
+        flags.select &:allowed
+      end
     end
   end
 end
